Adjust caption position to fit in the view bounds
This addresses #13 but does not yet comp[etely resolve it.

An issue remains around the initial layout where the caption size
can not be properly determined (it seems to ignore effects of CSS
even when forcing applyCss and performing a layout).

This manifests itself as a visual glitch on subequence repaints of
the caption (e.g. after a scene size change).