Improvements to annotation caption layout
The caption layout relied on measuring the size of the caption before
the initial display - but the size reports incorrectly (lacking the
effect of the associated CSS).

To ensure correct size, the caption node is added to a temporary
scene, then CSS and an initial layout are forced.

Subsequently adding the caption to the 'real' scene removes it from
the temporary one automatically.

Resolves #13.