Fix implementation of localization time window
The time window is like a "padding" used to extend either side of
the actual cued start and times for a localization.

This also means that any localization with a duration of zero will
display correctly with this padding both before and after the cued
start time.

Due to how the time ranges are calculated with this padding time
window it is undefined behaviour if the user makes a chage to the
confiurable time window setting while localizations are active.

This is because the start and times (plus this time window delta)
are used as keys in the range map.

Resolves #26.