Don't add an already existing localization
When adding new localizations it is desirable to immediately add
the dragged bounding box to the UI so there is no perceptible lag.

However this caused the same annotation to be added twice - once by
this immediate add, and again by the annotation manager when it
received a new time signal.

This is resolved simply by checking if the annotation was already
added before creating a new one.

Resolves #25.